Support both individual variables as well as files via deploy.sh
Presently --extra-vars for deploy.sh assumes that $2 is a file and
tacks on @ to the arg passed to ansible-playbook.  For some jobs it's
necessary to dynamically generate a variable
(e.g. $WORKSPACE\path\file.yaml) then pass it via -e.  This commit
changes the behavior to not always include '@'.

To pass a file, use '-e @file'.  This is consistent with -e handling
in ansible-playbook

Change-Id: Ibc7c6e59353bd7128756d3f4d2afda60726757d7